Validate.jl

Documentation for Validate.jl

Validate.Validation — Type

Validation

Struct to hold both the rules and the result of a confrontation.
Fields are

- rules: Validator-type holding all used rules
- out: Dict containing the result of applying each rule to a DataFrame

Validate.Validator — Type

Validator

Vector of Dicts to hold validation rules.

Rules have at least a `name` to uniquely identify a rule and an `expr`
holding the definition of a rule as a Julia expression. 
An expression is given in terms of equalities or inequalities between columns of a
DataFrame.  The result of the expression must be a boolean vector. The collection of
allowed operations is given in `OPERATORS`. Furthermore, the function `is.na` can be
used. It returns true for any element which is Julia-missing.

Validate.confront — Method

confront(df::AbstractDataFrame, rules::Validator)

Apply the given rules to a DataFrame and return the result

Returns a Validation-object with rules and out. The first contains the original rules, the second contains a dict with for every rule a vector with the result. The name of the rule is used as the key in this dict

Example

using DataFrames, Validate
df = DataFrame(speed=[1, 2, 3], distance=[10, 11, 100])
rules = Validate.read_rules("myrules.yaml")
result = Validate.confront(rules, df)

Validate.lacking — Method

lacking(df::AbstractDataFrame, confront_out::Validation)

Apply the validation results to the dataframe and only return rows that have one or more missing values for the rules

Validate.lacking — Method

lacking(df::AbstractDataFrame, rules::Validator)

Apply the rules to the dataframe and only return rows that have one or more missing values for the rules

Validate.read_rules — Method

read_rules(filename::String)

Read a Validate rule-file in yaml-format.

Returns a list of dicts where each element is a rule. The key expr contains the julia-ast of the rule

Examples

using Validate
myrules = Validate.read_rules("myrules.yaml")

Validate.satisfying — Function

satisfying(df::AbstractDataFrame, confront_out::Validation, include_missing=false)

Apply the validation results to the dataframe and only return rows that satisfy all validation rules. If include_missing is true, then expressions that evaluate to missing are also counted as satisfied. If false, then any expression with value missing is counted as violating

Returns a DataFrame

Validate.satisfying — Function

satisfying(df::AbstractDataFrame, rules::Validator, include_missing=false)

Apply the rules to the dataframe and only return rows that satisfy all validation rules. If include_missing is true, then expressions that evaluate to missing are also counted as satisfied. If false, then any expression with value missing is counted as violating

Returns a DataFrame

Validate.summary — Method

summary(confront_out::Validation)

Return a summary output of the validation as a DataFrame

Validate.violating — Function

violating(df::AbstractDataFrame, confront_out::Validation, include_missing=false)

Apply the validation results to the dataframe and only return rows that violate one of more rules. If include_missing is true, then expressions that evaluate to missing are also counted as satisfied. If false, then any expression with value missing is counted as violating

Returns a DataFrame

Validate.violating — Function

violating(df::AbstractDataFrame, rules::Validator, include_missing=false)

Apply the rules to the dataframe and only return rows that violate one of more validation rules. If include_missing is true, then expressions that evaluate to missing are also counted as satisfied. If false, then any expression with value missing is counted as violating

Returns a DataFrame